+++ source/Fable-UUID-Random.js
@@ -1,0 +1,23 @@
+/**
+* Random Byte Generator
+*
+* @license MIT
+*
+* @author Steven Velozo <steven@velozo.com>
+*/
+
+let libCrypto = require('crypto');
+
+class RandomBytes
+{
+	constructor()
+	{
+	}
+
+	generate()
+	{
+		return libCrypto.randomBytes(16);
+	}
+}
+
+module.exports = RandomBytes;