+++ CHANGELOG.md
@@ -6,3 +6,7 @@ Change Log
 
 #### 2015-04-03: Steven Velozo
 * Added some error handling
+
+#### 2019-04-19: Steven Velozo
+* Refactored to be an es6 module
+* Removed all external dependencies